requirements.txt
celery==4.1.0
Django==2.0.4
django-authtools==1.6.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-filter==1.1.0
django-model-utils==3.1.1
django-rest-swagger==2.1.2
django_google_authenticator==0.5
djangorestframework==3.8.2
djangorestframework-camel-case==1.0b1
docutils==0.14
ethereum==2.3.1
factory-boy==2.10.0
faker==0.8.12
firebase-admin==2.9.1
gevent==1.2.2
gunicorn==19.7.1
psycopg2-binary==2.7.4
web3==4.0.0
